pyproject.toml
[project]
name = "event-graph-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "accelerate>=1.7.0",
    "qwen-vl-utils>=0.0.10",
    "Pillow>=11.0.0",
    "tqdm>=4.67.0",
    "bitsandbytes>=0.48.1",
    "torch>=2.10.0",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
